moved the CSVParsing and JSONParsing to parsingMethods.js to reduce the code density in main.js
