+++ main.js
@@ -17,7 +17,7 @@ import htmlToCSV, {arrayToCsv, downloadCSVFile } from './js/functions/sideEffect
  
 /*============================={ parsers }=============================*/
 
-import { CSVParser, JSONParser, getObjKeys, getObjValues,fileParse } from './js/classes/CSVParser.js';
+import { CSVParsing, JSONParsing } from './js/parsingMethods.js';
 
 const form = document.querySelector("#getfile");
 const selectGroup = document.querySelector('#sort-select-group')
@@ -177,55 +177,6 @@ const papaparseParse = () => {
 
 };
 
-const CSVParsing = (data) => {
-  console.time('CSVParse')
-
-  const CSV = CSVParser(data.split('\n'));
-
-  const headerColumn = CSV[0];
-  const csvBody = CSV.slice(1)
-  const dataBody = removeUndefined(csvBody)
-
-  console.timeEnd('CSVParse')
-
-  return [headerColumn, dataBody]
-}
-
-const JSONParsing = (data) => {
-  console.time('JSON parsing');
-  let arrofObj = JSONParser(data);
-
-  const keys = getObjKeys(arrofObj[0])
-  const values = arrofObj.map(obj => {
-    return getObjValues(obj);
-  })
-
-  console.timeEnd('JSON parsing')
-  return [keys, values];
-}
-
-const parseHandler = (parser) => {
-  if(!selectedFile) return;
-  
-  var reader = new FileReader();
-
-  reader.readAsText(selectedFile)
-
-  reader.onload = async function (e) {
-    var data = e.target.result
-
-    const [headerColumn, dataBody] = parser(data);    
-
-    setDataPoints(headerColumn, select)
-    displayMethod(headerColumn, dataBody)   
-
-    updateBtn.onclick = () => {
-      onUpdate(headerColumn, dataBody)
-    }
-
-  }
-}
-
 const updateBtnWithoutClass = () => {
   Status.delegateOnclickEvent(
     {
@@ -265,6 +216,27 @@ const statusConfigOnDisplay = {
   enable: [clearBtn, updateBtn],
   disable: [submitBtn, displayBtn, inputFile],
   restrictSettings: parsingMethods
+}
+
+const parseHandler = (parser, cb) => {
+  if(!selectedFile) return;
+  
+  var reader = new FileReader();
+
+  reader.readAsText(selectedFile)
+
+  reader.onload = async function (e) {
+    var data = e.target.result
+
+    const [headerColumn, dataBody] = parser(data);    
+
+    setDataPoints(headerColumn, select)
+    displayMethod(headerColumn, dataBody)   
+
+    updateBtn.onclick = () => {
+      onUpdate(headerColumn, dataBody)
+    }
+  }
 }
 
 displayBtn.onclick = () => {
@@ -311,7 +283,7 @@ clearBtn.onclick = () => {
   csv.clear();
   Status.setStatus(statusConfigOnClear)
   select.innerHTML = '';
-
+  
 }
 
 const statusConfigOnUpdate = {
